Move assign-object.ts to common/helpers
