+++ src/common/exceptions/user-not-found.exception.ts
@@ -1,6 +1,6 @@
 import { BaseException } from "./base.exception";
 
-export class ShoesNotFoundException extends BaseException {
+export class UserNotFoundException extends BaseException {
     id = 102;
     statusCode = 404;
     message = 'User does not exist'
