+++ src/common/helpers/assign-object.ts
